+++ src/main/java/org/kdb/inside/brains/view/export/BinaryExportAction.java
@@ -2,7 +2,6 @@ package org.kdb.inside.brains.view.export;
 
 import com.intellij.openapi.fileChooser.FileChooserFactory;
 import com.intellij.openapi.fileChooser.FileSaverDescriptor;
-import com.intellij.openapi.fileChooser.FileSaverDialog;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFileWrapper;
@@ -24,8 +23,7 @@ public class BinaryExportAction extends AnExportAction<VirtualFileWrapper> {
     @Override
     protected VirtualFileWrapper getExportConfig(Project project, ExportDataProvider view) {
         final FileSaverDescriptor fileSaverDescriptor = new FileSaverDescriptor("Export to KDB Binary", "Exporting data into native KDB IPC format", "kib");
-        final FileSaverDialog saveFileDialog = FileChooserFactory.getInstance().createSaveFileDialog(fileSaverDescriptor, project);
-        return saveFileDialog.save("Table Result");
+        return FileChooserFactory.getInstance().createSaveFileDialog(fileSaverDescriptor, project).save(view.getExportName());
     }
 
     @Override