+++ src/main/java/org/kdb/inside/brains/view/export/CsvExportAction.java
@@ -3,7 +3,6 @@ package org.kdb.inside.brains.view.export;
 import com.google.common.primitives.Primitives;
 import com.intellij.openapi.fileChooser.FileChooserFactory;
 import com.intellij.openapi.fileChooser.FileSaverDescriptor;
-import com.intellij.openapi.fileChooser.FileSaverDialog;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFileWrapper;
@@ -22,10 +21,9 @@ public class CsvExportAction extends AnExportAction<VirtualFileWrapper> {
     }
 
     @Override
-    protected VirtualFileWrapper getExportConfig(Project project, ExportDataProvider dataProvider) {
+    protected VirtualFileWrapper getExportConfig(Project project, ExportDataProvider view) {
         final FileSaverDescriptor fileSaverDescriptor = new FileSaverDescriptor("Export to CSV", "Exporting data into tab separated file format", "csv");
-        final FileSaverDialog saveFileDialog = FileChooserFactory.getInstance().createSaveFileDialog(fileSaverDescriptor, project);
-        return saveFileDialog.save("Table Result");
+        return FileChooserFactory.getInstance().createSaveFileDialog(fileSaverDescriptor, project).save(view.getExportName());
     }
 
     @Override